apps/bluesky-poster/requirements.txt
atproto>=0.0.16
python-dotenv>=1.0.0
pyyaml>=6.0
questionary>=1.10.0
openai>=1.0.0
requests>=2.25.1